+++ CHANGES.md
@@ -5,6 +5,7 @@ Change Log
       edited from the Search by Badge Type screen with the proper permission)
     - Long names (and badge names) will be resized to fit on the badge (to a certain point)
     - Added automated install script for Ubuntu 16.04, updated installation instructions
+    - Final stripe colors for specialty badges set in base data
 
 - Version 0.9.6 (9/13/2016)
     - Changed error message to disappear automatically after a few seconds, instead of requiring a click